Re #213 - remove GDK usage before XInit
Apparently this fixes segfaults in our code for fixing segfaults.

This should be safe to do because if a package is built with X11 support
then the user needs libX11 installed, and the XInitThreads just
initialises some mutexes etc, so it won't crash with undefined symbols
and it shouldn't adversely affect someone running a Wayland server.